% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SYSDATA_OBJECTS.R
\name{SYSDATA_OBJECTS$errorcodes}
\alias{SYSDATA_OBJECTS$errorcodes}
\title{Error codes and messages of paramater checks executed}
\format{A data frame based on the data of inst/extdata/errorcodes.csv file within corresponding package.
\describe{
\item{\strong{errCode}}{ Field for number which identifies error entry}
\item{\strong{errMessage}}{ Field for message shown to user when error is detected by parameter check}
\item{\strong{description}}{ Field for description of error}
\item{\strong{solution}}{ Field for possible solution which hints at how to handle error }
\item{\strong{refFunction}}{ Field which references parameter check function where error entry belongs to}
}}
\description{
A dataset containtaining the error codes, messages, descriptions, solutions and references to function each belongs to.
}
\examples{
#USAGE in script within package COOP_GAME - e.g. selecting error entry where errCode is 1000:
SYSDATA_OBJECTS$errorcodes[SYSDATA_OBJECTS$errorcodes[,"errCode"]==1000,]
}
\seealso{
Other SYSDATA_OBJECTS: \code{\link{SYSDATA_OBJECTS$getSysdataErrorEntry}},
  \code{\link{SYSDATA_OBJECTS$getSysdataObject}},
  \code{\link{SYSDATA_OBJECTS}}
}

