% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaintenanceSysdata.R
\name{exportEmptyErrFile}
\alias{exportEmptyErrFile}
\title{exportEmptyErrFile}
\usage{
exportEmptyErrFile(csvErrFile = CSV_ERRFILE, packagePath = ".",
  errCodesStart = numeric(), errCodesEnd = numeric(), boolExtdata = TRUE)
}
\arguments{
\item{errCodesStart}{First number where error code range should start}

\item{errCodesEnd}{Last number where error code range should end}

\item{boolExtdata}{is boolean variable, when         'TRUE':  CSV file gets created in extdata directory of package
when default 'FALSE': CSV file get created according to referenced path by param 'csvErrFile'}

\item{file}{represents: in case boolExtdata is set   'FALSE': Path under which newly created csv file gets stored
'TRUE':  Name file gets stored in extdata directory of package}
}
\description{
exportEmptyErrFile creates new empty csv file with column names according to fields to be maintained for error codes
}
\examples{

#USECASE 1: C csv file within package
setwd("CoopGame");

#CoopGame  <-Working Directory
#└── inst
#|    └── extdata
#|        └── errorcodes.csv  <- csv file with errorcodes referenced by param fileErrCodes which will be created
#├── man
#├── man-roxygen
#├── R
#└── tests

exportEmptyErrFile(fileCSV="errorcodes.csv");

#or also, cause default value of csvErrFile:
exportEmptyErrFile();


#USECASE 2: Simple import of different objects when csv file is not within package

setwd("SVN_REPO");

#SVN_REPO  <-Working directory
#├── CommunicationGames
#├── CoopGame
#└── errorcodes.csv <- csv file with errorcodes referenced by param fileErrCodes

exportEmptyErrFile(fileCSV="./errorcodes.csv",packagePath="./CoopGame",boolExtdata=FALSE);

#or also, cause default value of csvErrFile:
exportEmptyErrFile(packagePath="./CoopGame",boolExtdata=FALSE);
}
\author{
Johannes Anwander
}
\seealso{
Other MaintenananceSysdata: \code{\link{SYSDATA_OBJECTS$getSysdataErrorEntry}},
  \code{\link{SYSDATA_OBJECTS}},
  \code{\link{exportErrSysdata}},
  \code{\link{exportPackageSysdata}},
  \code{\link{importErrDataFrame}},
  \code{\link{maintainCSVDataWithFix}}
}

