% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaintenanceSysdata.R
\name{exportErrSysdata}
\alias{exportErrSysdata}
\title{exportErrSysdata}
\usage{
exportErrSysdata(csvErrFile = CSV_ERRFILE, packagePath = ".",
  boolExtdata = TRUE, ...)
}
\arguments{
\item{csvErrFile}{represents the csv file containing the error codes and error messages}

\item{packagePath}{represents path to package 'sysdata.rda' gets written to.}

\item{boolExtdata}{is boolean variable, when         'TRUE':  param csvErrFile is path of csv file for importing error codes and messages
when default 'FALSE': param csvErrFile is name under which file is available in extdata directory of package}
}
\description{
exportErrSysdata creates 'sysdata.rda' file containing the error codes and error messages within the referenced package in the 'R' folder.
}
\section{Exported Functions}{

Functions to be exported on default when calling function exportErrSysdata.
\tabular{lll}{
\strong{Function Name}                   \tab  \strong{Short Description} \cr
\code{\link{getSysdataErrorEntry}}       \tab Returns entry for specified error code in errorcodes data.frame (list element of SYSDATA_OBJECTS) within sysdata  \cr
\code{\link{getSysdataObject}}           \tab Returns list object of SYSDATA_OBJECTS with specified object name. \cr
}
}
\examples{
#Creating 'sysdata.rda' file with error codes and messages
for package 'CoopGame' (analogue for 'CommunicationGames')

#USECASE 1 (recommended usage) - Working directory is set
#to folder of package 'CoopGame' (under development) and error codes
# and messages are stored in a CSV file within the same package in the folder 'inst/extdata/'

setwd("CoopGame");

#CoopGame  <-Working Directory
#├── inst
#│   └── extdata
#│       └── errorcodes.csv <-File referenced by csvErrFile
#├── man
#├── man-roxygen
#├── R
#└── tests

exportErrSysdata(csvErrFile="errorcodes.csv")

#or also, cause default value of csvErrFile:
exportErrSysdata()

#USECASE 2 - Working directory is not set to folder of package 'CoopGame' (under development) and CSV file with error codes and error message is external

setwd("SVN_REPO")

#SVN_REPO  <-Working directory
#├── CommunicationGames
#├── CoopGame
#└── errcodes.csv

exportErrSysdata(csvErrFile="./errorcodes.csv", packagePath="./CoopGame", boolExtdata=FALSE)
#or also, cause default value of csvErrFile:
exportErrSysdata(packagePath="./CoopGame", boolExtdata=FALSE)
}
\author{
Johannes Anwander
}
\seealso{
Other MaintenananceSysdata: \code{\link{SYSDATA_OBJECTS$getSysdataErrorEntry}},
  \code{\link{SYSDATA_OBJECTS}},
  \code{\link{exportEmptyErrFile}},
  \code{\link{exportPackageSysdata}},
  \code{\link{importErrDataFrame}},
  \code{\link{maintainCSVDataWithFix}}
}

