% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaintenanceSysdata.R
\name{exportPackageSysdata}
\alias{exportPackageSysdata}
\title{exportPackageSysdata}
\usage{
exportPackageSysdata(packagePath = ".", ...)
}
\arguments{
\item{packagePath}{represents path to package}

\item{...}{represents objects which get exported to 'sysdata.R'}
}
\description{
exportPackageSysdata creates 'sysdata.rda' file containing any objects for package
After creating file 'sysdata.rda' gets exported.
}
\examples{

#USECASE 1: Simple export of different objects to package when working directory is not set to package
setwd("CoopGame");

#CoopGame  <-Working Directory
#├── inst
#├── man
#├── man-roxygen
#├── R
#└── tests


a=5;
b=c;
funcDummy<-function(){print("HelloWorld!")}
createErrSysdataCoopGame(a,b,funcDummy)

#USECASE 2: Simple export of different objects to package when working directory is not set to package

setwd("SVN_REPO");

#SVN_REPO  <-Working directory
#├── CommunicationGames
#├── CoopGame
#└── errcodes.csv


a=5;
b=c;
funcDummy<-function(){print("HelloWorld!")}
errorcodes=createErrSysdataCoopGame(a,b,funcDummy, packagePath="./CoopGame");
exportPackageSysdata(a,b,funcDummy,errorcodes,packagePath="./CoopGame");
}
\author{
Johannes Anwander
}
\seealso{
Other MaintenananceSysdata: \code{\link{SYSDATA_OBJECTS$getSysdataErrorEntry}},
  \code{\link{SYSDATA_OBJECTS}},
  \code{\link{exportEmptyErrFile}},
  \code{\link{exportErrSysdata}},
  \code{\link{importErrDataFrame}},
  \code{\link{maintainCSVDataWithFix}}
}

