% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaintenanceSysdata.R
\name{MaintenananceSysdata}
\alias{MaintenananceSysdata}
\title{MaintenananceSysdata}
\description{
Functions to maintain sysdata.rda file of specified package
}
\section{Maintenance Functions}{

\tabular{lll}{
\strong{Function Name}                   \tab  \strong{Short Description} \cr
\code{\link{exportEmptyErrFile}}         \tab Creates and exports new empty csv file with column names according to fields to be maintained for error codes \cr
\code{\link{exportErrSysdata}}           \tab Creates and exports 'sysdata.rda' file containing the error codes and error messages as well as other objects. \cr
\code{\link{exportPackageSysdata}}       \tab Creates 'sysdata.rda' file containing any objects for package \cr
\code{\link{importErrDataFrame}}         \tab Imports errorcodes and corresponding attributes out of specified csv file \cr
\code{\link{maintainCSVDataWithFix}}     \tab Maintains any csv file with editor supplied by function fix \cr
}
}
\author{
Johannes Anwander
}

