% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SYSDATA_OBJECTS.R
\name{SYSDATA_OBJECTS}
\alias{SYSDATA_OBJECTS}
\title{SYSDATA_OBJECTS}
\format{A list of objects which are list items
\tabular{lll}{
\strong{List items:}                                \tab \strong{Short Description} \cr
\code{\link{SYSDATA_OBJECTS$errorcodes}}            \tab Dataset containtaining the error codes entries \cr
\code{\link{SYSDATA_OBJECTS$getSysdataErrorEntry}}  \tab Function to retrieve specified whole error entry or single attribute \cr
\code{\link{SYSDATA_OBJECTS$getSysdataObject}}      \tab Function to retrieve object out of sysdata.rda\cr
}}
\description{
A list containtaining the objects which are exported to sysdata.rda of corresponding package by function \code{\link{exportErrSysdata}}.
List items can be accessed for internal usage in script file of package.
}
\seealso{
Other MaintenananceSysdata: \code{\link{SYSDATA_OBJECTS$getSysdataErrorEntry}},
  \code{\link{exportEmptyErrFile}},
  \code{\link{exportErrSysdata}},
  \code{\link{exportPackageSysdata}},
  \code{\link{importErrDataFrame}},
  \code{\link{maintainCSVDataWithFix}}

Other SYSDATA_OBJECTS: \code{\link{SYSDATA_OBJECTS$errorcodes}},
  \code{\link{SYSDATA_OBJECTS$getSysdataErrorEntry}},
  \code{\link{SYSDATA_OBJECTS$getSysdataObject}}
}

