% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SYSDATA_OBJECTS.R
\name{SYSDATA_OBJECTS$fillParamCheckResult}
\alias{SYSDATA_OBJECTS$fillParamCheckResult}
\alias{fillParamCheckResult}
\title{fillParamCheckResult - fill error object with error code and message
#' @family SYSDATA_OBJECTS}
\usage{
fillParamCheckResult(paramCheckObject, errCode,
  packageName = getPackageName())
}
\arguments{
\item{paramCheckObject}{}
}
\description{
fillParamCheckResult retrieves error message out of data frame \code{\link{SYSDATA_OBJECTS$errorcodes}} by errCode and fills error object with error message and code.
}
\author{
Johannes Anwander
}

