% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SYSDATA_OBJECTS.R
\name{SYSDATA_OBJECTS$getSysdataErrorEntry}
\alias{SYSDATA_OBJECTS$getSysdataErrorEntry}
\alias{getSysdataErrorEntry}
\title{getSysdataErrorEntry}
\usage{
getSysdataErrorEntry(errCode, attributeName = "",
  packageName = getPackageName(),
  SYSDATA_OBJECTS = getFromNamespace("SYSDATA_OBJECTS", ns = packageName))
}
\arguments{
\item{errCode}{references specific error entry}

\item{SYSDATA_OBJECTS}{list containing objects which gets exported  sysdata.rda}
}
\description{
getSysdataErrorEntry returns entry for specified error code
}
\examples{

#USAGE in script within package COOP_GAME:
SYSDATA_OBJECTS$getSysdataErrorEntry(errCode = 1000, packageName = "CoopGame")

#Output:
#errCode                         errMessage description solution              refFunction
#1    1000 Game vector A is invalid as 'NULL'        <NA>     <NA> stopOnInvalidGameVectorA
}
\author{
Johannes Anwander
}
\seealso{
Other MaintenananceSysdata: \code{\link{SYSDATA_OBJECTS}},
  \code{\link{exportEmptyErrFile}},
  \code{\link{exportErrSysdata}},
  \code{\link{exportPackageSysdata}},
  \code{\link{importErrDataFrame}},
  \code{\link{maintainCSVDataWithFix}}

Other SYSDATA_OBJECTS: \code{\link{SYSDATA_OBJECTS$errorcodes}},
  \code{\link{SYSDATA_OBJECTS$getSysdataObject}},
  \code{\link{SYSDATA_OBJECTS}}
}

