% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SYSDATA_OBJECTS.R
\name{SYSDATA_OBJECTS$getSysdataObject}
\alias{SYSDATA_OBJECTS$getSysdataObject}
\alias{getSysdataObject}
\title{getSysdataObject}
\usage{
getSysdataObject(objectName = "", packageName = getPackageName(),
  SYSDATA_OBJECTS = getFromNamespace("SYSDATA_OBJECTS", ns = packageName))
}
\arguments{
\item{SYSDATA_OBJECTS}{list containing the objects which gets exported to sysdata.rda}

\item{errCode}{references specific error entry}
}
\description{
getSysdataObjects returns list object of SYSDATA_OBJECTS with specified object name.
}
\examples{

#USAGE in script within package COOP_GAME - e.g. retrieve error object
SYSDATA_OBJECTS$getSysdataObject(objectName = "errorcodes", packageName = getPackageName())
}
\author{
Johannes Anwander
}
\seealso{
Other SYSDATA_OBJECTS: \code{\link{SYSDATA_OBJECTS$errorcodes}},
  \code{\link{SYSDATA_OBJECTS$getSysdataErrorEntry}},
  \code{\link{SYSDATA_OBJECTS}}
}

